testing: Actually send an uncompressed packet in the ipv6/rw-compress-ikev2 scenario
The default of 56 bytes already exceeds the threshold of 90 bytes (8 bytes
ICMP + 40 bytes IPv6 = 104).  By reducing the size we make sure the
packet is not compressed (40 + 8 + 40 = 88).

This also fixes a strange failure of this scenario due to the recently
added post-test `ip xfrm state` check.  The kernel stores the used SAs
on the inbound skbuffs and since these are garbage collected it could
take a while until all references to an SA dissapear.  So while SAs
are not directly destroyed when we delete them, they are actually marked
as dead and won't show up in `ip xfrm state`.  However, that's not true
for the tunnel SA attached to the IPComp SA, which we don't explicitly
delete.  So when the last ping got compressed the skbuff had a reference
to the IPComp SA and therefore the tunnel SA.  If this skbuff was destroyed
after we ran our check the tunnel SA actually would still get reported and
the test case therefore failed.